backend_app/pyproject.toml
[project]
name = "backend_app"
version = "0.1.0"
authors = [
    { name = "Martin Pres", email = "martin.pres@prespol.cz" }
]
description = "Provides REST endpoints for visualizing collected data and controlling MQTT publishers"
requires-python = ">=3.10"
dependencies = [
    "paho-mqtt==2.1.0",
    "psycopg[binary,pool]==3.2.3",
    "pymongo==4.10.1",
    "djangorestframework == 3.15.2"
]

[project.scripts]
backend_app = "backend.manage:main"

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false
